Respect configured exchange name
Commit 4c92e62a removed the possibility of providing the exchange name in
the configuration, which means you need to run with exchange_declare set to
true and accept the fact that you will end up with one exchange per queue.

To handle the scenario with a named exchange that has multiple queues, it's
necessary to fetch the exchange name from the configuration.